Remove TCP and UDP reader wrappers
