+++ src/infrastructure/network/framing/tcp_encoder_test.go
@@ -1,4 +1,4 @@
-package chacha20
+package framing
 
 import (
 	"bytes"
@@ -6,10 +6,8 @@ import (
 	"testing"
 )
 
-// TestDefaultTCPEncoderEncode checks that Encode correctly writes the 4-byte length prefix.
 func TestDefaultTCPEncoderEncode(t *testing.T) {
 	encoder := NewDefaultTCPEncoder()
-	// Prepare a buffer: first 4 bytes reserved for length, followed by payload.
 	payload := []byte("Hello, World!")
 	buffer := make([]byte, 4+len(payload))
 	copy(buffer[4:], payload)
@@ -19,20 +17,17 @@ func TestDefaultTCPEncoderEncode(t *testing.T) {
 		t.Fatalf("Encode returned error: %v", err)
 	}
 
-	// The length prefix should equal the length of the payload.
 	expectedLength := uint32(len(payload))
 	gotLength := binary.BigEndian.Uint32(buffer[:4])
 	if gotLength != expectedLength {
 		t.Errorf("Expected length prefix %d, got %d", expectedLength, gotLength)
 	}
 
-	// Verify that the payload remains unchanged.
 	if !bytes.Equal(buffer[4:], payload) {
 		t.Errorf("Payload mismatch, expected %q, got %q", payload, buffer[4:])
 	}
 }
 
-// TestDefaultTCPEncoderDecode checks that Decode correctly populates a TCPPacket.
 func TestDefaultTCPEncoderDecode(t *testing.T) {
 	encoder := NewDefaultTCPEncoder()
 	data := []byte("Test packet data")
