+++ src/infrastructure/routing/client_routing/client_factory/worker_factory.go
@@ -8,8 +8,8 @@ import (
 	"time"
 	"tungo/application"
 	"tungo/infrastructure/PAL/client_configuration"
-	"tungo/infrastructure/cryptography/chacha20"
 	"tungo/infrastructure/network"
+	"tungo/infrastructure/network/framing"
 	"tungo/infrastructure/routing/client_routing/routing/tcp_chacha20"
 	"tungo/infrastructure/routing/client_routing/routing/udp_chacha20"
 	"tungo/infrastructure/settings"
@@ -36,12 +36,12 @@ func (w *WorkerFactory) CreateWorker(
 		}
 		transport := network.NewClientUDPAdapter(conn.(*net.UDPConn), deadline, deadline)
 		// tunHandler reads from tun and writes to transport
-		tunHandler := udp_chacha20.NewTunHandler(ctx, chacha20.NewUdpReader(tun), transport, crypto)
+		tunHandler := udp_chacha20.NewTunHandler(ctx, tun, transport, crypto)
 		// transportHandler reads from transport and writes to tun
 		transportHandler := udp_chacha20.NewTransportHandler(ctx, transport, tun, crypto)
 		return udp_chacha20.NewUdpWorker(transportHandler, tunHandler), nil
 	case settings.TCP:
-		encoder := chacha20.NewDefaultTCPEncoder()
+		encoder := framing.NewDefaultTCPEncoder()
 		tunHandler := tcp_chacha20.NewTunHandler(ctx, encoder, tun, conn, crypto)
 		transportHandler := tcp_chacha20.NewTransportHandler(ctx, conn, tun, crypto)
 		return tcp_chacha20.NewTcpTunWorker(ctx, tunHandler, transportHandler, crypto), nil
