+++ src/infrastructure/routing/client_routing/routing/tcp_chacha20/tun_handler.go
@@ -2,12 +2,13 @@ package tcp_chacha20
 
 import (
 	"context"
+	"encoding/binary"
 	"golang.org/x/crypto/chacha20poly1305"
 	"io"
 	"log"
 	"tungo/application"
-	"tungo/infrastructure/cryptography/chacha20"
 	"tungo/infrastructure/network"
+	"tungo/infrastructure/network/framing"
 )
 
 type TunHandler struct {
@@ -15,11 +16,11 @@ type TunHandler struct {
 	reader              io.Reader // abstraction over TUN device
 	writer              io.Writer // abstraction over transport
 	cryptographyService application.CryptographyService
-	encoder             chacha20.TCPEncoder
+	encoder             framing.TCPEncoder
 }
 
 func NewTunHandler(ctx context.Context,
-	encoder chacha20.TCPEncoder,
+	encoder framing.TCPEncoder,
 	reader io.Reader,
 	writer io.Writer,
 	cryptographyService application.CryptographyService) application.TunHandler {
@@ -33,7 +34,6 @@ func NewTunHandler(ctx context.Context,
 }
 
 func (t *TunHandler) HandleTun() error {
-	reader := chacha20.NewTcpReader(t.reader)
 	buffer := make([]byte, network.MaxPacketLengthBytes+4+chacha20poly1305.Overhead)
 
 	//passes anything from tun to chan
@@ -42,7 +42,7 @@ func (t *TunHandler) HandleTun() error {
 		case <-t.ctx.Done():
 			return nil
 		default:
-			n, err := reader.Read(buffer)
+			payloadLen, err := t.reader.Read(buffer[4:])
 			if err != nil {
 				if t.ctx.Err() != nil {
 					return nil
@@ -50,6 +50,8 @@ func (t *TunHandler) HandleTun() error {
 				log.Printf("failed to read from TUN: %v", err)
 				return err
 			}
+			binary.BigEndian.PutUint32(buffer[:4], uint32(payloadLen))
+			n := payloadLen
 
 			_, encryptErr := t.cryptographyService.Encrypt(buffer[4 : n+4])
 			if encryptErr != nil {
